Rename controller package and module
