+++ settings.gradle
@@ -27,7 +27,7 @@ include ':api:api-admins-config-v2'
 include ':api:api-agents-v4'
 include ':api:api-api-shared-v1'
 include ':api:api-artifact-store-config-v1'
-include ':api:api-authToken-v1'
+include ':api:api-accessToken-v1'
 include ':api:api-backups-v1'
 include ':api:api-base'
 include ':api:api-build_cause-v1'
